refactor: move BLANK_SPACES to constants.py
